+++ changelog.md
@@ -12,3 +12,4 @@ vX.X.X
 - Added statistics sub-tab, closes #79.
 - Added flashing TX/RX activity indicators, closes #80.
 - Added layout options for RX/TX tab, closes #81.
+- Changed the "UTF-8" decoding option to "ASCII", closes #83.